Fleshed out logistic regression, added textual explanation of approach
Initial dataset has proven to be extremely flawed, as it is both imbalanced and inaccurate. Will manually obtain fresh campaign finance info from FEC and electoral data from AP and limit dataset exclusively to on-ballot candidates before next commit. It is hoped that this will introduce some balance to the data, which are currently dominated by candidates with small campaign contributions and few votes and thus of little interest vis a vis the research hypothesis.